Remove EmulateOldDeliveryChannelProperties setting
Merge `WithOldDeliveryChannelProperties` tests with main test files
