+++ src/protagonist/API.Tests/Features/Queues/Validation/QueuePostValidatorTests.cs
@@ -10,11 +10,10 @@ namespace API.Tests.Features.Queues.Validation;
 
 public class QueuePostValidatorTests
 {
-    private QueuePostValidator GetSut(bool emulateOldDeliveryChannelProperties = false)
+    private QueuePostValidator GetSut()
     {
         var apiSettings = new ApiSettings
         {
-            EmulateOldDeliveryChannelProperties = emulateOldDeliveryChannelProperties,
             MaxBatchSize = 4
         };
         
@@ -205,23 +204,5 @@ public class QueuePostValidatorTests
         var model = new HydraCollection<Image> { Members = new[] { new Image { ThumbnailPolicy = "my-policy" } } };
         var result = sut.TestValidate(model);
         result.ShouldHaveValidationErrorFor("Members[0].ThumbnailPolicy");
-    }
-    
-    [Fact]
-    public void Member_ImageOptimisationPolicy_Allowed_WhenOldDeliveryChannelEmulationEnabled()
-    {
-        var sut = GetSut(true);
-        var model = new HydraCollection<Image> { Members = new[] { new Image { ImageOptimisationPolicy = "my-policy" } } };
-        var result = sut.TestValidate(model);
-        result.ShouldNotHaveValidationErrorFor("Members[0].ImageOptimisationPolicy");
-    }
-
-    [Fact]
-    public void Member_ThumbnailPolicy_Allowed_WhenOldDeliveryChannelEmulationEnabled()
-    {
-        var sut = GetSut(true);
-        var model = new HydraCollection<Image> { Members = new[] { new Image { ThumbnailPolicy = "my-policy" } } };
-        var result = sut.TestValidate(model);
-        result.ShouldNotHaveValidationErrorFor("Members[0].ThumbnailPolicy");
     }
 }
