+++ src/protagonist/API/Features/Queues/Validation/QueuePostValidator.cs
@@ -40,15 +40,11 @@ public class QueuePostValidator : AbstractValidator<HydraCollection<DLCS.HydraMo
             members.RuleFor(a => a.ModelId).NotEmpty().WithMessage("Asset Id cannot be empty");
             members.RuleFor(a => a.Space).NotEmpty().WithMessage("Space cannot be empty");
             
-            members.RuleFor(a => a.ImageOptimisationPolicy)
-                .Null()
-                .When(_ => !apiSettings.Value.EmulateOldDeliveryChannelProperties)
-                .WithMessage("'ImageOptimisationPolicy' is disabled");
+            members.RuleFor(a => a.ImageOptimisationPolicy).Null()
+                .WithMessage("'ImageOptimisationPolicy' can only be specified when legacy mode is enabled");
             
-            members.RuleFor(a => a.ThumbnailPolicy)
-                .Null()
-                .When(_ => !apiSettings.Value.EmulateOldDeliveryChannelProperties)
-                .WithMessage("'ThumbnailPolicy' is disabled");
+            members.RuleFor(a => a.ThumbnailPolicy).Null()
+                .WithMessage("'ThumbnailPolicy' can only be specified when legacy mode is enabled");
         });
     }
 }
